backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
supabase==1.0.3
python-dotenv==1.0.0